requirements.txt
# Web API
fastapi
uvicorn

pandas
numpy

logai; python_version < "3.13"

langchain
langchain-cohere
langchain-openai
langgraph
langsmith

sentence-transformers

google-generativeai

openai   # used as client for Grok endpoint

groq

cohere
datasets
ragas

scikit-learn

httpx

python-dotenv

pydantic

tqdm
python-dotenv
requests 
beautifulsoup4 
rank_bm25 
mcp
ragas
google-generativeai